拆分cluster client
