+++ pkg/client/redis/client.go
@@ -16,17 +16,9 @@ import (
 )
 
 type Client struct {
-	cluster *redis.ClusterClient
-	master  *redis.Client
-	slave   []*redis.Client
-	config  *Config
-}
-
-func (ins *Client) CmdOnCluster() *redis.ClusterClient {
-	if ins.cluster == nil {
-		ins.config.logger.Panic("redis:no cluster for "+ins.config.name, xlog.FieldExtMessage(ins.config))
-	}
-	return ins.cluster
+	master *redis.Client
+	slave  []*redis.Client
+	config *Config
 }
 
 func (ins *Client) CmdOnMaster() *redis.Client {
@@ -73,14 +65,6 @@ func (config *Config) Build() (*Client, error) {
 	var err error
 	if xdebug.IsDevelopmentMode() {
 		xdebug.PrettyJsonPrint("redis's config: "+config.name, config)
-	}
-
-	if len(config.Addr) > 0 {
-		ins.cluster, err = config.buildCluster()
-		if err != nil {
-			return ins, err
-		}
-		return ins, nil
 	}
 	if config.Master.Addr != "" {
 		addr, user, pass := getUsernameAndPassword(config.Master.Addr)
@@ -152,53 +136,4 @@ func (config *Config) build(addr, user, pass string) (*redis.Client, error) {
 		ClientStub: stubClient,
 	})
 	return stubClient, nil
-}
-
-func (config *Config) buildCluster() (*redis.ClusterClient, error) {
-	clusterClient := redis.NewClusterClient(&redis.ClusterOptions{
-		Addrs:        config.Addr,
-		Username:     config.Username,
-		Password:     config.Password,
-		MaxRetries:   config.MaxRetries,
-		DialTimeout:  config.DialTimeout,
-		ReadTimeout:  config.ReadTimeout,
-		WriteTimeout: config.WriteTimeout,
-		PoolSize:     config.PoolSize,
-		MinIdleConns: config.MinIdleConns,
-		IdleTimeout:  config.IdleTimeout,
-	})
-
-	for _, addr := range config.Addr {
-		clusterClient.AddHook(fixedInterceptor(config.name, addr, config, config.logger))
-		if config.EnableMetricInterceptor {
-			clusterClient.AddHook(metricInterceptor(config.name, addr, config, config.logger))
-		}
-		if config.Debug {
-			clusterClient.AddHook(debugInterceptor(config.name, addr, config, config.logger))
-		}
-		if config.EnableTraceInterceptor {
-			clusterClient.AddHook(traceInterceptor(config.name, addr, config, config.logger))
-		}
-		if config.EnableAccessLogInterceptor {
-			clusterClient.AddHook(accessInterceptor(config.name, addr, config, config.logger))
-		}
-
-		if config.EnableSentinel {
-			clusterClient.AddHook(sentinelInterceptor(config.name, addr, config, config.logger))
-		}
-	}
-
-	clusterClient.Ping(context.Background())
-	if err := clusterClient.Ping(context.Background()).Err(); err != nil {
-		if config.OnDialError == "panic" {
-			config.logger.Panic("redis cluster client start err: " + err.Error())
-		}
-		config.logger.Error("redis cluster client start err", xlog.FieldErr(err))
-		return nil, err
-	}
-
-	instances.Store(config.name, &storeRedis{
-		ClientCluster: clusterClient,
-	})
-	return clusterClient, nil
 }
