+++ pkg/client/redis/client_test.go
@@ -45,27 +45,3 @@ func Test_Stub(t *testing.T) {
 	})
 
 }
-
-func Test_Cluster(t *testing.T) {
-	config := DefaultConfig()
-	t.Run("should panic when addr nil", func(t *testing.T) {
-		var client *Client
-		defer func() {
-			if r := recover(); r != nil {
-				assert.Equal(t, "no master or slaves for default", r.(string))
-				assert.Nil(t, client)
-			}
-		}()
-		client = config.MustSingleton()
-		assert.Nil(t, client) // 不会执行到这里
-	})
-	t.Run("should not panic when dial err", func(t *testing.T) {
-		config.Addr = []string{"1.1.1.1"}
-		config.OnDialError = "error"
-		client, err := config.Build()
-		assert.NotNil(t, err)
-		assert.Nil(t, client.cluster)
-
-	})
-
-}
