+++ pkg/client/redis/config.go
@@ -126,6 +126,24 @@ func RawConfig(key string) *Config {
 	}
 
 	return config
+}
+
+func RawClusterConfig(key string) *Config {
+	var config = DefaultConfig()
+
+	if err := cfg.UnmarshalKey(key, &config, cfg.TagName("toml")); err != nil {
+		config.logger.Panic("unmarshal config:"+key, xlog.FieldErr(err), xlog.FieldName(key), xlog.FieldExtMessage(config))
+	}
+
+	if len(config.Addr) == 0 {
+		config.logger.Panic("no cluster addr set:"+key, xlog.FieldName(key), xlog.FieldExtMessage(config))
+	}
+	config.name = key
+	if xdebug.IsDevelopmentMode() {
+		xdebug.PrettyJsonPrint(key, config)
+	}
+
+	return config
 
 }
 func getUsernameAndPassword(addr string) (realAddr string, username, password string) {
